feat: Persist shelf selections across pages.
When a user selects a bookshelf for a book from the search page,
it will not be reflected within the shelf picker select. Additionally,
if a book is already on a shelf the shelf picker select will be prepopulated
with the shelf it is on.

When a user selects a shelf for a particular book, they are able to immediately
go back to the Bookshelf view and see the book on the page.